# docs/openapi-scoreboard.yaml
openapi: 3.1.0
info:
  title: falco-ctf scoreboard
  version: "2"
  description: |
    The scoreboard binary's COMPLETE HTTP surface (ADR-0005).

    Subject of this document
    ------------------------
    **Every route `cmd/scoreboard`'s mux registers — not "the JSON API".**
    That includes the HTML shells (`GET /`, `GET /portal`), the vendored
    stylesheet, `/healthz` and `/metrics`. Non-JSON routes get a thin entry
    (summary + statuses + audience, no schema); they are here so the
    documented route set can be compared to the registered route set as an
    EXACT SET (ADR-0005 Verification). A spec whose scope is "the API only"
    needs a maintained exclusion list, and the exclusion list is exactly how
    this document decayed to 50% coverage before ADR-0005.

    One spec per service, audience per operation
    --------------------------------------------
    Participant and operator routes live in ONE document because they are
    served by ONE binary off ONE mux, and OpenAPI expresses authorization
    per-operation. Splitting by audience would spread one route set over two
    files, where a route can go missing from both. Read the audience off
    `x-ctf-audience` / `x-ctf-authz` on each operation, never off the URL
    prefix.

    Per-operation extensions (ALL are mandatory on every operation)
    --------------------------------------------------------------
      x-ctf-audience          participant | operator | internal | infra
...

...

...
        `status` only.
...
servers:
...
paths:
...
  /api/challenges/{cid}/submit-detect:
    post:
      summary: Submit a Falco condition for a detect-type challenge
      operationId: submitDetect
      tags: [participant]
      x-ctf-audience: participant
      x-ctf-authz: self-or-admin-write
      x-ctf-origin-guard: true
      x-ctf-collector-forward: false
      x-ctf-rate-limit: per-IP 1 req/s burst 10 + global in-flight grader cap (5)
      description: |
        Grade a participant-authored Falco `condition:`. It is wrapped into a
        fixed rule and replayed against the mission's evasion + benign
        captures; `pass = fires on evasion AND no fire on benign`. The reference
        condition and the capture contents are NEVER returned — only fire
        counts.

        ONE caller (the portal's Grade button), so the origin guard is safe to
        apply here: the collector does not forward this route. On an
        auth-proxied host the claimed body `user` must be the proven identity
        (or an admin). Past the global in-flight cap the request is REJECTED
        with 429, never queued (a per-submission grader Job is the net-new DoS
        lever). With no runner configured the route answers 503 (feature off),
        not 404/500.
      parameters:
        - $ref: "#/components/parameters/Cid"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitDetectRequest" }
      responses:
        "200":
          description: graded — branch on `status`, use the fire counts as feedback
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SubmitDetectVerdict" }
        "400":
          description: not a detect challenge, missing/invalid `user`, missing `condition`, or condition too large
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: origin guard denied, or caller is neither `{user}` nor an admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: unknown challenge id
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: |
            per-IP rate limit OR the global in-flight grader cap — two
            different writers (internal/scoreboard/ratelimit and this
            handler), both JSON since Issue #159 unified ratelimit onto
            httpx.WriteJSON.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: grader infra error (fail-closed — never a silent solve)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503":
          description: detect grading is not enabled (no runner wired)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/users/{user}/challenges/{cid}/hints/{idx}:
    post:
      summary: Reveal one progressive hint (in order, idempotent)
      operationId: openHint
      tags: [participant]
      x-ctf-audience: participant
      x-ctf-authz: self-or-admin-write
      x-ctf-origin-guard: true
      x-ctf-collector-forward: false
      x-ctf-rate-limit: per-IP 1 req/s burst 10
      description: |
        Hints open in order: opening `idx` N requires N-1 already open (409
        otherwise), so participants approach the answer progressively instead of
        jumping to the last hint. Idempotent — re-opening returns the text again
        with `newly: false`.

        Revealing costs points (the per-index schedule lives in the scoring
        layer; `Journey.detail.hints.penalty` prices the NEXT reveal). Hint copy
        comes from `journey.yaml` and by convention carries no flag values
        (conventions I10).

        Origin-guarded for the same reason as steps/check, plus a second one:
        the 409/200 distinction is a cross-user oracle if an unauthenticated
        caller can drive it.
      parameters:
        - $ref: "#/components/parameters/User"
        - $ref: "#/components/parameters/Cid"
        - in: path
          name: idx
          required: true
          schema: { type: integer, minimum: 1 }
          description: 1-based hint index, bounded by the mission's hint count
      responses:
        "200":
          description: hint revealed (or already open)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OpenHintResult" }
        "400":
          description: invalid user or hint index out of range
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: origin guard denied, or caller is neither `{user}` nor an admin
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: unknown challenge id, or the mission has no hints
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: previous hint not open yet (in-order reveal)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: rate limited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: could not record the hint view
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/board/threads:
    get:
      summary: List the QA Board threads the caller may see
      operationId: boardListThreads
      tags: [participant]
      x-ctf-audience: participant
      x-ctf-authz: authenticated
      x-ctf-origin-guard: false
      x-ctf-collector-forward: false
      x-ctf-rate-limit: none
      description: |
        app#292 (destination model, replaces P25's per-user QA ticket chat
        wholesale): every `audience: all` thread (visible to every
        authenticated participant) PLUS the caller's own `audience: admin`
        threads (private, visible only to their own author and admins) —
        never another participant's private thread. `authenticated`, not
        `self-or-admin`: this route carries no `{user}` path segment to
        match against: the caller's OWN identity (from
        `X-Auth-Request-Email`) is implicitly "self" and there is nothing
        else to authorize. Hidden/deleted threads are omitted (moderation
        state is admin-only visibility — see `GET /api/admin/board/threads`
        for the full queue).
      responses:
        "200":
          description: thread summaries, pinned-first then most-recently-active first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardList" }
        "403":
          description: missing/blank identity header
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: could not list board threads
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    post:
      summary: Open a new QA Board thread
      operationId: boardCreateThread
      tags: [participant]
      x-ctf-audience: participant
      x-ctf-authz: authenticated
      x-ctf-origin-guard: true
      x-ctf-collector-forward: false
      x-ctf-rate-limit: identity 1 req/10s burst 3 (shared with boardAppendMessage)
      description: |
        `author` / `author_role` are ALWAYS server-set (`participant` / the
        caller's own derived username) — the request schema below has no
        such fields at all, so there is nothing for a participant to
        smuggle a role claim through; any such keys in a raw request body
        are silently ignored (same discipline P25's createQuestion
        enforced, security-engineer finding 1).

        `audience` selects visibility: `all` (public — any authenticated
        participant may read and like it) or `admin` (private — visible
        only to its own author and admins). Fail-closed: any value OTHER
        than exactly `admin` or `all` (including an absent/empty field) is
        treated as `admin` — a malformed or omitted audience never
        accidentally becomes public.

        The opening `body` (and every other free-text board write) is
        rejected 400 if it matches the shape `FALCO{...}` — a fairness
        control, not a real-flag comparison (the board has no access to any
        challenge's actual flag value; see the handler's own doc for why
        blanket shape-rejection, not value comparison, is the correct gate
        here).

        Origin-guarded: reached ONLY from the portal's browser fetch, never
        the collector's forward allowlist (`docs/openapi-collector.yaml`) —
        same reasoning ADR-0006 gave for declining to add P25's QA there.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateThreadRequest" }
      responses:
        "200":
          description: thread created (its one message is the participant's own opening body)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardThread" }
        "400":
          description: malformed body, body empty/over the length cap, or body matches the flag shape
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "403":
          description: origin guard denied, or missing/blank identity header
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: rate limited
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "500":
          description: could not create the thread
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  parameters:
...
    Cid:
      in: path
      name: cid
      required: true
      schema: { type: string }
      description: challenge id / catalog key (e.g. `03-stealth-read`)
...
  schemas:
    Error:
      type: object
      description: |
        The ONLY body shape for a non-2xx JSON response. `error` is meant to be
        a stable, operator-facing message; the implementation still leaks
        decoder / SQLite driver text on some paths (Issue #113) — that is a
        defect against this schema, not an alternative contract.
      required: [error]
      properties:
        error: { type: string }
...
    SubmitDetectRequest:
      type: object
      required: [user, condition]
      properties:
        user: { type: string }
        condition:
          type: string
          description: the Falco condition to grade (<= 4 KiB)
    SubmitDetectVerdict:
      type: object
      description: |
        200 verdict for `POST /api/challenges/{cid}/submit-detect`. `status` is
        the branch key; `user` / `display_name` appear only on `solved`, and the
        fire counts are absent on `invalid` (the condition never compiled).
        `reason` appears on `invalid`/`missed`/`false-positive` but NOT on
        `solved` (the solved branch has nothing to explain).
      required: [status, solved]
      properties:
        status:
          type: string
          enum: [invalid, missed, false-positive, solved]
        solved: { type: boolean }
        evasion_fires: { type: integer }
        benign_fires: { type: integer }
        user: { type: string }
        display_name: { type: string }
        reason: { type: string }
...
    BoardList:
      type: object
      required: [threads]
      properties:
        threads:
          type: array
          items: { $ref: "#/components/schemas/BoardSummary" }
...